flickit-assessment-users/src/main/java/org/flickit/assessment/users/application/service/space/UpdateSpaceService.java
package org.flickit.assessment.users.application.service.space;

import lombok.RequiredArgsConstructor;
import org.flickit.assessment.users.application.port.in.space.UpdateSpaceUseCase;
import org.flickit.assessment.users.application.port.out.space.LoadSpacePort;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
@Transactional
@RequiredArgsConstructor
public class UpdateSpaceService implements UpdateSpaceUseCase {

    private final LoadSpacePort loadSpacePort;

    @Override
    public Result updateSpace(Param param) {
        loadSpacePort.loadSpace(param.getId());
        return null;
    }
}
